training: moving huggingface trainer to job module
